+++ scripts/run_collect_and_compile.sh
@@ -11,8 +11,8 @@ set -e
 if [ -z "$ORA_BENCH_RUN_CX_ORACLE_PYTHON" ]; then
     export ORA_BENCH_RUN_CX_ORACLE_PYTHON=true
 fi
-if [ -z "$ORA_BENCH_RUN_EXPOSED_KOTLIN" ]; then
-    export ORA_BENCH_RUN_EXPOSED_KOTLIN=true
+if [ -z "$ORA_BENCH_RUN_JDBC_KOTLIN" ]; then
+    export ORA_BENCH_RUN_JDBC_KOTLIN=true
 fi
 if [ -z "$ORA_BENCH_RUN_GODROR_GO" ]; then
     export ORA_BENCH_RUN_GODROR_GO=true
@@ -67,7 +67,7 @@ echo "RUN_GLOBAL_JAMDB           : $RUN_GLOBAL_JAMDB"
 echo "RUN_GLOBAL_NON_JAMDB       : $RUN_GLOBAL_NON_JAMDB"
 echo "--------------------------------------------------------------------------------"
 echo "RUN_CX_ORACLE_PYTHON       : $ORA_BENCH_RUN_CX_ORACLE_PYTHON"
-echo "RUN_EXPOSED_KOTLIN         : $ORA_BENCH_RUN_EXPOSED_KOTLIN"
+echo "RUN_JDBC_KOTLIN         : $ORA_BENCH_RUN_JDBC_KOTLIN"
 echo "RUN_GODROR_GO              : $ORA_BENCH_RUN_GODROR_GO"
 echo "RUN_JAMDB_ORACLE_ERLANG    : $ORA_BENCH_RUN_JAMDB_ORACLE_ERLANG"
 echo "RUN_JDBC_JAVA              : $ORA_BENCH_RUN_JDBC_JAVA"
@@ -182,7 +182,7 @@ if [ "$RUN_GLOBAL_NON_JAMDB" = "true" ]; then
         echo "Setup Go - End   ===========================================================" 
     fi    
 
-    if [ "$ORA_BENCH_RUN_EXPOSED_KOTLIN" == "true" ]; then
+    if [ "$ORA_BENCH_RUN_JDBC_KOTLIN" == "true" ]; then
         echo "Setup Kotlin - Start =======================================================" 
         if ! ./src_kotlin/scripts/run_gradle.sh; then
             exit 255
@@ -192,7 +192,7 @@ if [ "$RUN_GLOBAL_NON_JAMDB" = "true" ]; then
 
     if [ "$ORA_BENCH_RUN_CX_ORACLE_PYTHON" == "true" ]; then
         echo "Setup Python - Start =======================================================" 
-        if ! java -cp "priv/libs/*" ch.konnexions.orabench.OraBench setup_python; then
+        if ! java -cp "$ORA_BENCH_JAVA_CLASSPATH" ch.konnexions.orabench.OraBench setup_python; then
             exit 255
         fi
         echo "Setup Python - End   ======================================================="